requirements.txt
# FastAPI + server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database (async SQLite for Phase 1, Postgres for Phase 2)
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
# asyncpg>=0.29.0  # Uncomment for Phase 2 Postgres

# Pydantic settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client for Modal proxy (Phase 2)
httpx>=0.26.0

# JWT handling
python-jose[cryptography]>=3.3.0

# Utils
python-dotenv>=1.0.0
